Extract tags service layer for MCP reuse and code deduplication (#552)
Create src/server/services/tags.ts with listTags, createTag, updateTag,
and deleteTag functions following the established service layer pattern.
Refactor tags router to delegate to these service functions instead of
containing inline business logic. Add MCP tools for tag management
(list_tags, create_tag, update_tag, delete_tag).
